# Skip dummy node creation for autograd engine when there is a single input and place on correct queue