--- a/torch/csrc/autograd/engine.cpp
+++ b/torch/csrc/autograd/engine.cpp
@@ -880,10 +880,11 @@ auto Engine::execute(const edge_list& roots,
     InputBuffer input_buffer(roots.at(0).function->num_inputs());
     auto input = inputs.at(0);
 
+    const auto input_stream = InputMetadata(input).stream();
     const auto opt_next_stream = roots.at(0).function->stream(c10::DeviceType::CUDA);
     input_buffer.add(roots.at(0).input_nr,
                       std::move(input),
-                      InputMetadata(input).stream(),
+                      input_stream,
                       opt_next_stream);
 
     execute_with_graph_task(graph_task, graph_root, std::move(input_buffer));