--- a/torch/csrc/autograd/python_engine.h
+++ b/torch/csrc/autograd/python_engine.h
@@ -28,7 +28,8 @@ struct PythonEngine : public Engine {
 
   std::shared_ptr<at::ivalue::Future> execute_with_graph_task(
       const std::shared_ptr<GraphTask>& graph_task,
-      std::shared_ptr<Node> graph_root) override;
+      std::shared_ptr<Node> graph_root,
+      variable_list inputs = {}) override;
 
   std::unique_ptr<AnomalyMetadata> make_anomaly_metadata() override;
   private: